GUI/backend/requirements.txt
fastapi
uvicorn
websockets
cassandra-driver
